step 26

+++ src/main/java/com/ll/sb20231114/domain/article/article/controller/ArticleController.java
@@ -43,6 +43,8 @@ public class ArticleController {
 
     @GetMapping("/article/write")
     String showWrite() {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         return "article/article/write";
     }
 
@@ -56,6 +58,8 @@ public class ArticleController {
 
     @PostMapping("/article/write")
     String write(@Valid WriteForm writeForm) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         Article article = articleService.write(writeForm.title, writeForm.body);
 
         return rq.redirect("/article/list", "%d번 게시물 생성되었습니다.".formatted(article.getId()));
@@ -63,6 +67,8 @@ public class ArticleController {
 
     @GetMapping("/article/modify/{id}")
     String showModify(Model model, @PathVariable long id) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         Article article = articleService.findById(id).get();
 
         model.addAttribute("article", article);
@@ -80,6 +86,8 @@ public class ArticleController {
 
     @PostMapping("/article/modify/{id}")
     String modify(@PathVariable long id, @Valid ModifyForm modifyForm) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         articleService.modify(id, modifyForm.title, modifyForm.body);
 
         return rq.redirect("/article/list", "%d번 게시물 수정되었습니다.".formatted(id));
@@ -87,6 +95,8 @@ public class ArticleController {
 
     @GetMapping("/article/delete/{id}")
     String delete(@PathVariable long id) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         articleService.delete(id);
 
         return rq.redirect("/article/list", "%d번 게시물 삭제되었습니다.".formatted(id));
